requirements.txt
annotated-types==0.6.0
anyio==3.7.0
astroid==2.15.5
click==8.1.3
dill==0.3.6
fastapi==0.110.0
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
numpy==1.24.3
pandas==2.0.2
platformdirs==3.5.3
pyarrow==12.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pylance==0.4.21
pylint==2.17.4
pysondb-v2==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
structlog==24.1.0
tomlkit==0.11.8
typing_extensions==4.10.0
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
